+++ projects/ids-enterprise-ng/src/lib/dropdown/soho-dropdown.component.ts
@@ -12,7 +12,9 @@ import {
   Optional,
   AfterViewChecked,
   ChangeDetectionStrategy,
-  ChangeDetectorRef
+  ChangeDetectorRef,
+  SimpleChange,
+  OnChanges
 } from '@angular/core';
 
 import {
@@ -28,7 +30,7 @@ import {
   template: '<ng-content></ng-content>',
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class SohoDropDownComponent implements AfterViewInit, AfterViewChecked, OnDestroy {
+export class SohoDropDownComponent implements AfterViewInit, AfterViewChecked, OnDestroy, OnChanges {
   /**
    * Used to provide unnamed controls with a unique id.
    */
@@ -137,7 +139,6 @@ export class SohoDropDownComponent implements AfterViewInit, AfterViewChecked, O
   }
 
   public get moveSelectedToTop(): boolean | undefined {
-    // eslint-disable-next-line import/no-deprecated
     return this.options.moveSelectedToTop;
   }
 
@@ -449,6 +450,9 @@ export class SohoDropDownComponent implements AfterViewInit, AfterViewChecked, O
     return this.options.multiple;
   }
 
+  @HostBinding('attr.readonly') @Input()
+  readonly: boolean | undefined;
+
   /**
    * Creates an instance of SohoDropDownComponent.
    *
@@ -512,6 +516,18 @@ export class SohoDropDownComponent implements AfterViewInit, AfterViewChecked, O
         setTimeout(() => this.updated());
         this.runUpdatedOnCheck = false;
       });
+    }
+  }
+
+  ngOnChanges(changes: any) {
+    if (changes['readonly']) {
+      this.ngZone.runOutsideAngular(() => {
+        if (this.readonly) {
+          this.dropdown?.readonly();
+        } else {
+          this.dropdown?.enable();
+        }
+      })
     }
   }
 
@@ -626,16 +642,6 @@ export class SohoDropDownComponent implements AfterViewInit, AfterViewChecked, O
     if (this.dropdown) {
       if (value) {
         this.ngZone.runOutsideAngular(() => this.dropdown?.disable());
-      } else {
-        this.ngZone.runOutsideAngular(() => this.dropdown?.enable());
-      }
-    }
-  }
-
-  @Input() set readonly(value: boolean) {
-    if (this.dropdown) {
-      if (value) {
-        this.ngZone.runOutsideAngular(() => this.dropdown?.readonly());
       } else {
         this.ngZone.runOutsideAngular(() => this.dropdown?.enable());
       }